Repository introduced

+++ back/src/pkg/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID           uint   `gorm:"primary_key;auto_increment;uniqueIndex;not null'" json:"ID"`
+	ID           int64  `gorm:"primary_key;auto_increment;uniqueIndex;not null'" json:"ID"`
 	Username     string `gorm:"uniqueIndex;notnull"`
 	PasswordHash string `gorm:"notnull"`
 }
